fix warnings / tidy code
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51270 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/util/clipper/OutRec.java
@@ -18,8 +18,8 @@ class OutRec {
 		}
 		double a = 0;
 		do {
-			a = a + (double) (op.prev.getPt().getX() + op.getPt().getX())
-					* (double) (op.prev.getPt().getY() - op.getPt().getY());
+			a = a + (op.prev.getPt().getX() + op.getPt().getX())
+					* (op.prev.getPt().getY() - op.getPt().getY());
 			op = op.next;
 		} while (op != pts);
 		return a * 0.5;
